Minor refinements to UI to improve look and usability. Also added sass support for bootstrap customization.

+++ checkbook-ui/src/components/checkbook/CheckbookListItem.js
@@ -4,24 +4,27 @@ import { deleteCheckbook } from "../../lib/api"
 
 const CheckbookListItem = (props) => {
     const navigate = useNavigate();
+    let formattedBalance = (Math.round(props.currentBalance * 100) / 100).toFixed(2);
     const handleDelete = () => {
-        deleteCheckbook(props.id).then(() => {
-            navigate('/checkbook');
-        });
+        if(window.confirm("Are you sure you want to delete this checkbook and all of it's transactions?")){
+            deleteCheckbook(props.id).then(() => {
+                navigate('/checkbook');
+            });
+        }
     };
 
     return (
-        <div className="card mt-1">
+        <div className="card shadow mt-3">
             <div className="card-header">
                 {props.name}
-                <i className="bi bi-trash float-end" onClick={handleDelete}></i>
+                <i className="bi bi-trash float-end" style={{'cursor':'pointer'}} onClick={handleDelete}></i>
             </div>
             <div className="card-body">
                 <h5 className="card-title">{props.name}</h5>
-                <p className="card-text">${props.currentBalance}</p>
+                <p className="card-text">${formattedBalance}</p>
                 <div className="row justify-content-center">
                     <div className="col-sm-6">
-                        <Link to={`/checkbook/${props.id}/transactions`} className="btn btn-primary">View Transactions</Link>
+                        <Link to={`/checkbook/${props.id}/transactions`} className="btn btn-success">View Transactions</Link>
                     </div>
                 </div>
             </div>
