+++ checkbook-ui/src/components/transactions/TransactionListItem.js
@@ -8,16 +8,18 @@ const TransactionListItem = (props) => {
     const isWithdrawal = props.type === 'W';
     //const itemClass =  isWithdrawal ? 'list-group-item-danger' : 'list-group-item-success';
     const handleDelete = () => {
-        deleteTransaction(props.id).then(() => {
-            navigate(`/checkbook/${params.checkbookId}/transactions`)
-        });
+        if(window.confirm("Are you sure you want to delete this transaction?")) {
+            deleteTransaction(props.id).then(() => {
+                navigate(`/checkbook/${params.checkbookId}/transactions`)
+            });
+        }
     };
 
     return (
-        <div className="card mt-1" style={{'backgroundColor': isWithdrawal ? 'tomato':'lightgreen'}}>
+        <div className="card shadow mt-3" style={{'backgroundColor': isWithdrawal ? '#ff3333':'lightgreen'}}>
             <div className="card-header">
                 <h5 className="mb-1 d-inline">{isWithdrawal ? 'Withdrawal' : 'Deposit'}</h5>
-                <i className="bi bi-trash float-end" onClick={handleDelete}></i>
+                <i className="bi bi-trash float-end" style={{'cursor':'pointer'}} onClick={handleDelete}></i>
             </div>
             <div className="card-body">
                 <h5 className="card-title">{props.description}</h5>
